Update pre-requisites.md
Changed step 2 in order to pull the downgraded version of fermipy. It was: sudo docker pull fermipy/fermipy (to download the most recent one). Doing the downgrade, we avoid the error `ImportError: No module named ipaddress` that was happening in step 4.iii.

Change step 4.iii so the downgraded version of fermipy is launched.

Excluded the "NOTE" in step 4.iii that explained how to downgrade fermipy only if necessary. The NOTE was:
---
**NOTE**

If you get an error message that ends with `ImportError: No module named ipaddress` when running the command above, then you need to downgrade your Fermipy version with the command

    sudo docker pull fermipy/fermipy:11-05-02

Then use the following command to launch jupyter notebook:

    sudo docker run -it --rm -p 8888:8888 -v $PWD:/workdir -w /workdir fermipy/fermipy:11-05-02

---

This changes avoids the `ImportError: No module named ipaddress`. It also guarantees that everybody is using the same version of fermipy and no compatibility problem can result of the usage of different versions.

OBS: the downgraded version of fermipy that we are using above (version 11-05-02) is not compatible with 4FGL and with P8R3 (and also doesn't contain the most recents galactic and isotropic diffuse models). The tutorial must take this into account (using 3FGL and P8R2, for instance).